move code to lib folder
